TestScript 실행을 TeshShell class로부터 분리
